Fix construction of the command word.
Some of the bit shifting math did not work because an int is 16 bits
and some of the bit shifting math ended up shifting the data out of
this range.

+++ schedule.c
@@ -11,8 +11,8 @@ void setPattern(Schedule* schedule, int index, int duration, bool red, bool yell
     }
 }
 
-int cycleTime(Schedule* schedule) {
-    int sum=0;
+uint16_t cycleTime(Schedule* schedule) {
+    uint16_t sum=0;
     int i;
 
     for(i = 0 ; i < MAX_PATTERNS ; i++) {
